Add badge color segments template support Unify gauge functions
