+++ src/const.ts
@@ -1,4 +1,5 @@
 export const DEFAULT_MIN = 0;
 export const DEFAULT_MAX = 100;
+export const MAX_ANGLE = 270;
 
 export const NUMBER_ENTITY_DOMAINS = ["sensor", "number", "counter", "input_number"];
