+++ src/utils/gauge.ts
@@ -1,3 +1,9 @@
+import { TemplateResult, svg } from "lit";
+import { MAX_ANGLE } from "../const";
+import type { SegmentsConfig } from "../card/type";
+import { rgbToHex } from "./color";
+import { interpolateRgb } from "d3-interpolate";
+
 type Vector = [number, number];
 type Matrix = [Vector, Vector];
 
@@ -67,3 +73,107 @@ export const svgArc = (options: ArcOptions) => {
     eY,
   ].join(" ");
 };
+
+export const strokeDashArc = (from: number, to: number, min: number, max: number, radius: number): [string, string] => {
+  const start = valueToPercentage(from, min, max);
+  const end = valueToPercentage(to, min, max);
+
+  const track = (radius * 2 * Math.PI * MAX_ANGLE) / 360;
+  const arc = Math.max((end - start) * track, 0);
+  const arcOffset = start * track - 0.5;
+
+  const strokeDasharray = `${arc} ${track - arc}`;
+  const strokeDashOffset = `-${arcOffset}`;
+  return [strokeDasharray, strokeDashOffset];
+}
+
+export const getAngle = (value: number, min: number, max: number) => {
+  return valueToPercentage(isNaN(value) ? min : value, min, max) * MAX_ANGLE;
+}
+
+export const valueToPercentage = (value: number, min: number, max: number) => {
+  return (clamp(value, min, max) - min) / (max - min);
+}
+
+export function renderSegments(segments: SegmentsConfig[], min: number, max: number, radius: number, smooth_segments: boolean | undefined): TemplateResult[] {
+  if (segments) {
+    let sortedSegments = [...segments].sort((a, b) => Number(a.from) - Number(b.from));
+
+    if (smooth_segments) {
+      let gradient: string = "";
+      sortedSegments.map((segment, index) => {
+        const angle = getAngle(Number(segment.from), min, max) + 45;
+        const color = typeof segment.color === "object" ? rgbToHex(segment.color) : segment.color;
+        gradient += `${color} ${angle}deg${index != sortedSegments.length - 1 ? "," : ""}`;
+      });
+      return [svg`
+        <foreignObject x="-50" y="-50" width="100%" height="100%" transform="rotate(45)">
+          <div style="width: 100px; height: 100px; background-image: conic-gradient(${gradient})">
+          </div>
+        </foreignObject>
+      `];
+    } else {
+      return [...sortedSegments].map((segment, index) => {
+        let roundEnd: TemplateResult | undefined;
+        const startAngle = index === 0 ? 0 : getAngle(Number(segment.from), min, max);
+        const angle = index === sortedSegments.length - 1 ? MAX_ANGLE : getAngle(Number(sortedSegments[index + 1].from), min, max);
+        const color = typeof segment.color === "object" ? rgbToHex(segment.color) : segment.color;
+        const segmentPath = svgArc({
+          x: 0,
+          y: 0,
+          start: startAngle,
+          end: angle,
+          r: radius,
+        });
+
+        if (index === 0 || index === sortedSegments.length - 1) {
+          const endPath = svgArc({
+            x: 0,
+            y: 0,
+            start: index === 0 ? 0 : MAX_ANGLE,
+            end: index === 0 ? 0 : MAX_ANGLE,
+            r: radius,
+          });
+          roundEnd = svg`
+          <path
+            class="segment"
+            stroke=${color}
+            d=${endPath}
+            stroke-linecap="round"
+          />`;
+        }
+
+        return svg`${roundEnd}
+          <path
+            class="segment"
+            stroke=${color}
+            d=${segmentPath}
+          />`;
+      });
+    }
+  }
+  return [];
+}
+
+export function computeSegments(numberState: number, segments: SegmentsConfig[] | undefined, smooth_segments: boolean | undefined): string | undefined {
+  if (segments) {
+    let sortedSegments = [...segments].sort((a, b) => Number(a.from) - Number(b.from));
+    
+    for (let i = 0; i < sortedSegments.length; i++) {
+      let segment = sortedSegments[i];
+      if (segment && (numberState >= Number(segment.from) || i === 0) &&
+        (i + 1 == sortedSegments?.length || numberState < Number(sortedSegments![i + 1].from))) {
+          if (smooth_segments) {
+            const color = typeof segment.color === "object" ? rgbToHex(segment.color) : segment.color;
+            const nextSegment = sortedSegments[i + 1] ? sortedSegments[i + 1] : segment;
+            const nextColor = typeof nextSegment.color === "object" ? rgbToHex(nextSegment.color) : nextSegment.color;
+            return interpolateRgb(color, nextColor)(valueToPercentage(numberState, Number(segment.from), Number(nextSegment.from)));
+          } else {
+            const color = typeof segment.color === "object" ? rgbToHex(segment.color) : segment.color;
+            return color;
+          }
+      }
+    }
+  }
+  return undefined;
+}
